fix: minor 수정 (#3)

+++ src/main/java/com/snowballer/api/dto/QuestionDto.java
@@ -4,6 +4,6 @@ import lombok.Getter;
 
 @Getter
 public class QuestionDto {
-	private int id;
-	private int answerId;
+	private Long id;
+	private Long answerId;
 }
